Move the disease analyzer out of WIP.
This necessitated some changes; it's in a disease2/ hierarchy for
some reason (presumably legacy), and was named British-style.
We're going to aim for consistent American spelling, so this "fixes"
that and also gets rid of the meaningless (to us) versioning.

These are placed directly in the maps, so those had to be updated
as well.