Move monkey dispenser out of WIP.
The last of the easy ones.  This was part of the virus2/disease2
"library," for some reason, but has nothing to do with any of the
other stuff; it's just a straight-up monkey dispenser.  So I got
rid of the disease2/ hierarchy.  Note that right now this only
appears on antiqua.