There were two completely parellel filing cabinet definitions.
I took the more interesting one that uses folders, and replaced the
existing boring cabinet with it.   It's cooler because of the folders
(which still need to be scattered about the maps), and because when
someone puts things in or takes things out, the drawer pops open.

The other filing_cabinet was renamed properly and moved to unused/