Incorporate atmospherics map debugging properly.
It's a client function, but we don't want it on by default, so it's
now controlled by the Debug2 variable rather than having to be turned
off and on at compile time.  Yes, Debug2 is a terrible, terrible name,
but it's used all over the place; it will be changed in a separate
commit.