Axe a bunch of unused #FILE_DIRs.
BYOND's IDE is idiotic and puts every single directory it sees into
a #include FILE_DIR.  Those are only used for .dmi files and sounds,
and most of our directory tree is code, not art and sound.  Since
we are no longer using the IDE's .dme anyway, we can clean this
garbage out.

Turns out that it seems to speed up compilation in Linux by about
ten seconds because it's not checking all of these directories that
will never have .dmi files.  Oh, BYOND.