Unmix code/datum/mixed.dm.
Most of the stuff in there involved /datum/data (yeah... it hurts
me too), so move the file there.  The definition for powernets was
randomly jammed inside too, though, so make a new file for that.
There was also a weird /datum/debug defined, but nothing outside
of unused/ used it, so toss that.