The great plasma -> phoron changeover.
This is a massive, massive commit, touching a whole lot of different
files, that changes all user-visible (and most code-visible) instances
of 'plasma' to 'phoron'.  The only things not changed were any
obvious icon references.  Amazingly, 'plasma.dmi' was the only live
file with plasma in its name.

Note that I mentioned obvious icon references.  I have no doubt that,
with my non-full understanding of BYOND and SS13, I changed something
that will break some graphic somewhere.  So, everyone, be on the
lookout for such things.