The great space carp die-off.
Space carp are silly, but not fun-silly.  Remove them from the game.
Note that this removes quite a few recipes that could easily be
added back (using regular fish rather than space carp); this is left
as an exercise for the reader.

Because the carp spawnpoints are also used for space ninja spawning,
they're now called 'ninjaspawn' rather than 'carpspawn.'