Clean up/tweak/rearrange the ion storm AI laws.
They were made into a list in a weird way (via code rather than
just making it a single variable), there were a bunch of rules
that could potentially screw individual crewmembers (which is
not much fun), and some potentially offensive ones.  There was
also one that explicitly referred to the Exodus, which has been
changed to referring to LOCATION_NAME.  Plus the usual typos and
poor grammar.  This cleans all that up.