Move all wizard spells into proc_holder/spell.dm.
There's a lot of files with them, and then one file with a whole lot,
and wizards are bad anyway.  Now we have a single file we can nuke
to get rid of a lot of their code when we get around to it.

Compiled successfully, but I won't lie--if this broke wizards, I'm
not gonna cry much.