Added script that expands resource file names to full paths.
By running this script on the code base, the compile time of the
entire project goes way down because Dream Maker no longer has
to search through a bunch of directories for every resource file.
On one system the compile time went from 44s to 36s. On another it
went from 3m34s to 1ms32s.

The script itself must be run from the root directory with the .dme
file as "tools/expand_filedir_paths.py"