Ran the expand_filedir_paths.py on the code base.
This touched all *.dm files and the single *.dmm map file.
The .dme file still has all the original FILE_DIR defines
in it, but these should no longer have an effect on compile
speed because now every resource file will be matched on the
very first FILE_DIR for ".". Eventually we can remove the
rest of them as well.