Move code/proc/global.dm to code/proc.dm.
Although this move pains me, it's the one place where we
(intentionally) broke the object tree -> filesystem mapping.  I
naively thought that there wouldn't be a whole lot of base-level
definitions.  There are lots, so code/ will be littered with .dm
files whether I like it or not... meaning that there's no reason
for this asymmetry.