An experimental tool to automatically cleanup .dm files
This tool will automatically detect the start of top-level class
definitions and will redirect them to the proper output .dm
file for each class path. The tool should be run from the base
directory with the .dme file as follows with FILEPATH replaced
with the input file one wishes to cleanup.

states -s default -f tools/code_cleanup.st < FILEPATH.dm | sh

The output will be a shell script with multiple "here documents"
that when executed, will append the various class definitions to
the proper output files. In practice, one should examine the
output of the shell script and make any manual corrections as
needed.

This tool is written for the "states" command which is part of
GNU Enscript. This mini-language is similar to awk, but the
regex patterns can match anywhere, unlike the line oriented
nature of awk. Essentially, it's a scripted lexical analyzer.