No meteor #1: Get rid of meteorhit().
The current meteor-strike implementation is pretty terrible, and
the terrible starts with this: everyone has to implement their own
"meteorhit" function, which presumably does something interesting
in the case of being struck by a meteor.  In most cases this was
just a copy of the equivalent blob-hit function, but in any case
the vast majority of objects in the game defaulted to the base-level
implementation... which did nothing.  This is a lot of complexity
(and an ugly function that gets attached to everything) for very
little gain.  So it's gone.