Fix pointers to the wrong lighting.dmi.
Wojo's script did not take into account the fact that multiple .dmis
can have the same name.  This didn't bite us for mecha, because the
files were actually identical (?!), but it picked the wrong one for
lighting, which made all lighting fixtures invisible.  Amazingly, no
one complained to us about this; it was first noticed when I, as a
captain on a test run, found a particular hallway really dark but
could not find any fixture with a broken bulb to replace.