Added a wrapper code_cleanup.sh shell script
The wrapper shell script makes it easy to apply the code cleanup tool
on entire directories at a time. Instructions for use are included in
the comment block at the beginning of the shell script.

Also fixed a small issue with code_cleanup.st having a double // in
the output filenames and with getting confused by a substring match
on "proc" when it was used as part of a longer name such as:

/foo/bar/proc_holder/stuff/proc/foo()