Fix bug in code_cleanup.st with handling of global vars and consts
The code is supposed to remove the /var/global/ or /var/const/ from
the variable name and replace it with a tab because that's how
all the globals in code/var/global.dm are organized. Unfortunately
the regex was leaving a / at the beginning of the tab indented
var name.