Move all unused soundfiles to unused/.
I realized after the last commit (of the unused .dmis) that I could
actually verify that these were unused by checking the sizes of the
Daedalus compiled files (particularly the .rsc).  This commit does
not change the size of that file, so I am (somewhat more) confident
that they are genuinely unused.