Further improvements to the SQLite interface
Started work on the .dm script wrapper around the native library.
This wrapper is modelled after dbcore.dm which seems to be the
standard library for using MySQL with BYOND.

My version aims to be as close as possible in API so that it can
be a drop in replacement. I'm using my own class names though
because I intend to eventually release it as a standalone BYOND
library and it shouldn't class with the existing MySQL library.