Improved decoding of row data in SQLite query
The dmsqlite.cpp wrapper now returns pointer values to databases
and statements. This allows for any number of statements to be
executing against the same database file.

The encoding of row data was modified slightly to make it easier
to process in the .dm script. The script itself is now able to
split a row of data into a BYOND list.