De-WIP central atmospherics' circuit board.
This is the first entry in what will inevitably end up a pretty big
file, as all circuit boards are fundamentally identical.  This leads
me to believe that the right thing to do is to have a single circuit
board type that can change its description depending on what it's
for rather than cluttering up the object tree... but for the time
being this will do.