Move the dynamic maploader to unused/ for now.
We may eventually want to re-add this functionality, but Daed doesn't
use it at the moment, and when we /do/ re-add it we're going to want
to look really, really, REALLY hard at the code to make sure it can't
totally trash the system.  As such, there's no current uses for it on
our side, so to unused/ it goes.