Remove useless tags on cables in the map.
I first noticed this because the tag was "Streight" (typos ahoy!).
When I dug further, I discovered that this is used... precisely
nowhere.  Presumably it's a consequence of bad copy/paste, or the
use of an external map editor that doesn't know how to spell.  There
is also a "90Curve" tag used on some (but not all) 90-degree turns
in the cabling... and all the more elaborate set-ups don't have tags
at all.

Of course, these tags come at the cost of extra memory, tracking,
etc. but more importantly they're just straight up line noise.