/obj/machinery/power/solar -> solar_panel.
This is one of those cases where, even though the first name is
not very ambiguous, it needs to change anyway.  That's because there
are /other/ objects named /obj/machinery/power/solar_XXX, which makes
finding instances of this versus those a pain in the butt.  It's also
a better name, but I probably wouldn't have bothered if not for the
existence of this name conflict.