Remove four ion_act() definitions.
If you've read my rambling extended entries, you'll know that I think
that the proliferation of xxx_act() functions is terrible design.
Instead, there should be one "quirky damage" function, that takes as
an optional parameter the type of damage, so that 90% of the stuff
can be the same with slight tweaks to make it specific if you so choose.

Apparently at some point there was something called ion damage.  At
least, I'm guessing so from the name of this function.  Anyway, in the
act of cleaning out the SMES code, I noticed this function... and then
saw that there was not a single caller left in the codebase.  Sigh.  So
it's gone.