Rename obj/machinery/light -> light_fixture.
After some contemplation, I realized that these light related things
I'm working on now should be better named.  The machinery is a fixture,
and the "weapon" is a lamp, as that's the generic name for "bit you
plug into a fixture so as to produce nifty glows."

This change affects way less than I thought it would, to be honest,
although it /does/ change roughly a billionty lines in the map file.
It also pointed out a rather stupid (but harmless) error in the
electrical storm code that I will be fixing momentarily.