Uncomplicate electrical storm lists.
For some reason, all of the electrical storm lists were of the
light fixture type.  I tried to make them their actual types,
but that... did not seem to work.  Which was troubling, but
lists are untyped in BYOND anyhow, so why complicate matters?
They're just plain lists now.  (Yes, this is FINALLY the bug
I noticed.)