Rename obj/item/weapon/light -> lamp.
For similar reasons to the light_fixture renaming I just did, I wanted
to rename the actual things-you-stick-in-fixtures to something other
than "light."  Wikipedia and my old electrical textbooks agree that
these should be called lamps.  Yes, I agree that there is ambiguity
there with the /fixture/ commonly called a lamp... except that that
now has the full name of obj/machinery/light_fixture/lamp, which...
is not ambiguous at all!  Yay for sane naming schemes.