New global function, get_reagent_color_mix().
When I looked at changing the name of GetColors(), I saw that all
of its callsites were cut-and-paste copies of the same functionality,
that of admixing reagents to come up with a new colour.  Being me,
I realized that this should be a single function rather than a whole
bunch of copies.  And, of course, I figured I could improve on the
original implementation.  Thus, get_reagent_color_mix().  This is
currently unused, but not for long.