Use get_reagent_color_mix() in the right places.
Now that I've written a function that replaces a bunch of copy-pasted
code, it's time to replace that copy-pasted code with calls to the
function.  This was an easy fix, and it caught a spot of total BS:
assigning a value to itself does not change its value to a "slightly
darker version."  Ugh.