Completely rewritten (and bugfixed) color mixer.
The previous implementation was explicitly wrong in how it handled
multiple reagents; specifically, the further down the list a reagent
was, the more of an influence it had on the final colour.  This is
wrong.  I rewrote the code completely to get an equal mix of all
involved reagents, leaving the quick-escape for the standard single
case.