Added SQLite support in the codebase.
I've changed all DBConnection and DBQuery references to their
equivalent SQLite class types, and so far everything appears
to be working. The library.dm code has not been tested yet
because it is separately disabled via a #define, and that needs
to be moved out of the file anyway.

The Linux build of dmsqlite.so works just fine, but the Windows
version deadlocks unless I disable all multi-thread (and hence
all mutexes) in the SQLite code. This may be a bug in my latest
version of Mingw, or maybe it's a side effect of statically
linking the GCC runtimes.

Either way, further proof that I need to compile the Windows
version with Microsoft compilers.

To enable SQLite, uncomment the ENABLE_STAT_TRACKING line in
config/dbconfig.txt and manually create the database schema with:
sqlite3 data/daedalus.db < SQL/daedalus_schema.sql