Fix argument order in text2fifo() and start using it
My previous text2fifo() had the arguments in the wrong order.
Since text2file() expects the text first and then the file name,
text2fifo() should be coded the same way to be consistent.

I replaced all the recently added instances of text2file() that
write to a FIFO with text2fifo() calls.