Better error handling logic
Rather than returning a fully-formed array when working and a completely
different array( 'Error' => 1 ) when broken, the Byond class now
precreateѕ a Null-fulled array with a new member, 'Status'.  If this is
set to 'Error', we add an 'Error' field which contains a description of
what went wrong, embedded into the full array.  This means we don't have
to do any hacky array_key_exists BS.

Updated PngGenerator to work with this new logic.