Removed Tajarans.
They're a cool idea, but the way they're implemented is a mess.
Tajaran-specific code is cut-and-pasted all over the place.  DG
and I both definitely want other species at some point... but
the code needs to be made sane first.  As I described to him,
"Strip down, then build up."